Add a --fd-gui CLI flag for a parent process to tell jackline when it has focus
- example: https://github.com/infinity0/jackline-gtk/blob/949e9db3/jackline-gtk#L175

This allows better notification behaviour. Previously, jackline would not
behave intuitively when receiving a message on the currently-active contact.
Specifically (a) if focused, this would cause "has notifications" even though
I am reading the messages right now, and I can only clear these by switching
away from then back to this contact and (b) if unfocused, then these are not
cleared even if I switch focus back to jackline.

The fundamental cause was that jackline used "active contact" as an heuristic
to detect when messages are read (to clear notifications for that contact) but
this heuristic is not correct when jackline itself does not have UI focus.

This does make jackline more aware of the UI than is necessary, but I believe
this is the simplest option that is still quite well-separated. The alternative
would be to tell the UI the following information:

1. for each message received, which jid/"random bucket" it was received on
2. when the user makes each jid/"random bucket" active in jackline

since this information plus (3) gui_has_focus is exactly the information needed
to correctly determine notifications. I decided that this commit was a simpler
solution since jackline already contained logic to deal with 1, 2.