Revert "regression: notify even if contact is active (and no gui_focus is installed)"
This reverts commit 3d910af43908d231368217a4cccd0b0b38b7a12b.

This commit isn't necessary, this case is covered with d6dcd27a - the difference
is that here the notifications are of the form `connected notify_contact
clear_all_notifications` (1) to indicate that the notification was immediately
cleared, because the contact is active.

For comparison, the old behaviour (which this commit restores) is to emit
`connected_notifications notify_contact` - however there are no remaining
notifications so the first argument is a bit misleading.

Notification callbacks do need to be re-written for (1), sorry if that wasn't
clear - one would check that the 3rd argument (2nd in --fd-nfy) is
`notify_contact` rather than checking that the 2nd argument (1st in --fd-nfy)
ends with `_notifications`.