redefine the semantics of an "active_session":
the common case is one or multiple online sessions, sort by priority.

in case there is no such online presence, we used to take the first offline one
if and only if we had no presence subscription of the buddy.  otherwise, there
was no active session.

this lead to trouble in cases where a (slightly broken) s2s connection does not
transport presence updates, since we weren't ever able to establish an OTR
connection.

this also blew up if a buddy went invisible (you know, `/status offline` and
`/disconnect` are different).

in slack land, we already "fixed" the receiving side, to deal with messages from
bare jid.

there may be unwanted behaviour changes: if a contact legible goes offline,
we'll still be able to access their last otr context.  but i guess that is ok
(and/or even improves the current situation).