switch to displaying dates as soon as the day changes
Previously we switched to displaying dates as soon as 24h passed. This
is however potentially confusing for users:

04-08 16:00 -O> Some message
17:00:00 -O> Some other message

Was "Some other message" sent today or yesterday? To answer this
question the user must:

* Check the current time
* If it is after  17:00:00, it was today
* If it is before 17:00:00, it was yesterday

With the new behavior, we prefix the times with a date as soon as we
cross the day border. This way it requires no awareness of the current
time to determine on which day a message was received: All messages
without a date are today, all messages with a date are yesterday or
earlier.

Additionally the rendering of old messages changes only once per day.
Previously it was kind of irritating if you were using jackline around
the same time on multiple days without clearing the history since every
once in a while a single message further up changes because its
timestamp is suddenly rendered differently.